Add resuming option for dypolychord with MPI

+++ lenstronomy/Sampling/Samplers/polychord_sampler.py
@@ -4,6 +4,7 @@ import os
 import shutil
 import numpy as np
 import sys
+import copy
 
 from lenstronomy.Sampling.Samplers.base_nested_sampler import NestedSampler
 import lenstronomy.Util.sampling_util as utils
@@ -20,7 +21,8 @@ class DyPolyChordSampler(NestedSampler):
 
     def __init__(self, likelihood_module, prior_type='uniform', 
                  prior_means=None, prior_sigmas=None, width_scale=1, sigma_scale=1,
-                 output_dir=None, output_basename='-', seed_increment=1,
+                 output_dir=None, output_basename='-',
+                 resume_dyn_run=False,
                  polychord_settings={},
                  remove_output_dir=False, use_mpi=False): #, num_mpi_procs=1):
         """
@@ -33,6 +35,7 @@ class DyPolyChordSampler(NestedSampler):
         :param output_dir: name of the folder that will contain output files
         :param output_basename: prefix for output files
         :param seed_increment: seed increment for dypolychord RNG with MPI. Check dypolychord documentation for details.
+        :param resume_dyn_run: if True, previous resume files will not be deleted so that previous run can be resumed
         :param polychord_settings: settings dictionary to send to pypolychord. Check dypolychord documentation for details.
         :param remove_output_dir: remove the output_dir folder after completion
         :param use_mpi: Use MPI computing if `True`
@@ -61,16 +64,16 @@ class DyPolyChordSampler(NestedSampler):
         else:
             self._comm = None
 
-        if self._is_master:
-            if os.path.exists(self._output_dir):
-                shutil.rmtree(self._output_dir, ignore_errors=True)
-            os.mkdir(self._output_dir)
+        if not resume_dyn_run:
+            if self._is_master:
+                if os.path.exists(self._output_dir):
+                    shutil.rmtree(self._output_dir, ignore_errors=True)
+                os.mkdir(self._output_dir)
 
         self._output_basename = output_basename
-        self._seed_increment = seed_increment
-        self._settings = polychord_settings
-        self._settings['file_root'] = self._output_basename,
-        self._settings['base_dir'] =  self._output_dir
+        self._settings = copy.deepcopy(polychord_settings)
+        self._settings['file_root'] = self._output_basename
+        self._settings['base_dir'] = self._output_dir
 
         if self._all_installed:
             # create the dyPolyChord callable object
@@ -134,7 +137,6 @@ class DyPolyChordSampler(NestedSampler):
             # dynamic_goal = 0 for evidence-only, 1 for posterior-only
 
             self._dyPolyChord.run_dypolychord(self._sampler, dynamic_goal,
-                                              seed_increment=self._seed_increment,
                                               settings_dict_in=self._settings,
                                               comm=self._comm, **kwargs_run)
